requirements.txt
prometheus_client
pyowm
python_nest
